+++ src/Runtime/Main/Characters/Player2D.cs
@@ -57,8 +57,6 @@ namespace MestreTramador
 
             Force = 300.0f;
             Speed = 5.0f;
-
-            JumpMode = JumpModes.DoubleJump;
         }
 
         /// <summary>
@@ -67,6 +65,17 @@ namespace MestreTramador
         protected override void OnUpdate()
         {
             E.ListenOnUpdate(this);
+        }
+
+        /// <summary>
+        ///     On Reset, all base sets are made and then
+        ///     a new Jump Mode is defined.
+        /// </summary>
+        protected override void OnReset()
+        {
+            base.OnReset();
+
+            JumpMode = JumpModes.DoubleJump;
         }
 
         /// <summary>